Ensuring picking new card from DB deson't pick card that was already COTD

+++ model/card_of_the_day.go
@@ -2,7 +2,7 @@ package model
 
 type CardOfTheDay struct {
 	Date    string `bson:"date" json:"date"`
-	Version uint8  `bson:"version" json:"version"`
+	Version int    `bson:"version" json:"version"`
 	CardID  string `bson:"cardID" json:"-"`
 	Card    Card   `bson:"-" json:"card"`
 }
